Unfixed the fix for drowning, changed settings xml.
The fix for drowning in battle caused other bugs (such as change weapon not working properly). This means that status-effects and drowning still happens in real time.

Config changed such that you can decide what mob types can enter turn-based battle with a player.

Old config will be renamed and a new config will be generated in its place.

+++ burnedkirby/TurnBasedMinecraft/gui/BattleGui.java
@@ -308,7 +308,7 @@ public class BattleGui extends GuiScreen {
 			break;
 		case 4: //Change weapon menu
 			info[0] = "Pick your weapon!";
-			for(int i=0; i < 9; i++)
+			for(short i=0; i < 9; i++)
 			{
 				buttonList.add(new ItemSelectionButton(6, width/2 - 88 + i * 20, height - 19, 16, 16, "", i));
 			}
@@ -361,7 +361,7 @@ public class BattleGui extends GuiScreen {
 			turnChoiceSent = true;
 			break;
 		case 6: //change weapon
-			int itemStackID = ((ItemSelectionButton)button).getItemStackID();
+			short itemStackID = ((ItemSelectionButton)button).getItemStackID();
 			Minecraft.getMinecraft().thePlayer.inventory.currentItem = itemStackID;
 			
 			player.type = Type.CHANGE_WEAPON;
